refactor Student to contain CourseChoice which has course, lectureSection, and tutSection
