+++ src/test/java/TestStudentCourseSorting.java
@@ -1,10 +1,9 @@
 import org.example.coursegetter.entities.Course;
-import org.example.requisitechecker.usecases.internal.BracketDealer;
-import org.example.studentdata.Student;
+import org.example.coursegetter.entities.CourseChoice;
+import org.example.studentdata.entities.Student;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -13,7 +12,7 @@ import java.util.Map;
 import static java.util.Map.entry;
 
 public class TestStudentCourseSorting {
-    private Course returnCourseFromName(String courseName){
+    private CourseChoice returnCourseChoiceFromName(String courseName){
         Map<String, Object> courseInfo = Map.ofEntries(
                 entry("orgName", "orgName"),
                 entry("code", courseName),
@@ -30,7 +29,9 @@ public class TestStudentCourseSorting {
                 entry("corequisite", "corequisite"),
                 entry("meetings", new HashMap<String, Object>())
         );
-        return new Course(courseInfo);
+        Course course = new Course(courseInfo);
+
+        return new CourseChoice(course, "LEC100", "TUT100");
     }
 
     @Test(timeout = 50)
@@ -38,9 +39,9 @@ public class TestStudentCourseSorting {
         String[] courseList = {"CSC210Y1", "CSC110Y1", "CSC310Y1", "CSC410Y1"};
         String[] expectedCourseList = {"CSC110Y1", "CSC210Y1", "CSC310Y1", "CSC410Y1"};
 
-        ArrayList<Course> courseArrayList = new ArrayList<>();
+        ArrayList<CourseChoice> courseArrayList = new ArrayList<>();
         for (String courseName: courseList) {
-            courseArrayList.add(returnCourseFromName(courseName));
+            courseArrayList.add(returnCourseChoiceFromName(courseName));
         }
 
         Student student = new Student();
@@ -49,11 +50,11 @@ public class TestStudentCourseSorting {
         student.plannedYCourses = new ArrayList<>();
         student.sortAllCourseLists();
 
-        List<Course> actualCourseArrayList = student.plannedFCourses;
+        List<CourseChoice> actualCourseArrayList = student.plannedFCourses;
         String[] actualCourseList = new String[actualCourseArrayList.size()];
 
         for (int i = 0; i < actualCourseArrayList.size(); i++) {
-            actualCourseList[i] = actualCourseArrayList.get(i).getCode();
+            actualCourseList[i] = actualCourseArrayList.get(i).course.getCode();
         }
 
         Assert.assertEquals(expectedCourseList, actualCourseList);
@@ -64,9 +65,9 @@ public class TestStudentCourseSorting {
         String[] courseList = {"CSC150Y1", "CSC130Y1", "CSC120Y1", "CSC110Y1", "CSC140Y1"};
         String[] expectedCourseList = {"CSC110Y1", "CSC120Y1", "CSC130Y1", "CSC140Y1", "CSC150Y1"};
 
-        ArrayList<Course> courseArrayList = new ArrayList<>();
+        ArrayList<CourseChoice> courseArrayList = new ArrayList<>();
         for (String courseName: courseList) {
-            courseArrayList.add(returnCourseFromName(courseName));
+            courseArrayList.add(returnCourseChoiceFromName(courseName));
         }
 
         Student student = new Student();
@@ -75,11 +76,11 @@ public class TestStudentCourseSorting {
         student.plannedYCourses = new ArrayList<>();
         student.sortAllCourseLists();
 
-        List<Course> actualCourseArrayList = student.plannedFCourses;
+        List<CourseChoice> actualCourseArrayList = student.plannedFCourses;
         String[] actualCourseList = new String[actualCourseArrayList.size()];
 
         for (int i = 0; i < actualCourseArrayList.size(); i++) {
-            actualCourseList[i] = actualCourseArrayList.get(i).getCode();
+            actualCourseList[i] = actualCourseArrayList.get(i).course.getCode();
         }
 
         Assert.assertEquals(expectedCourseList, actualCourseList);
@@ -90,9 +91,9 @@ public class TestStudentCourseSorting {
         String[] courseList = {"CSC100Y8", "CSC100Y3", "CSC100Y5", "CSC100Y4", "CSC100Y1", "CSC100Y7", "CSC100Y0", "CSC100Y2", "CSC100Y6", "CSC100Y9"};
         String[] expectedCourseList = {"CSC100Y1", "CSC100Y0", "CSC100Y5", "CSC100Y3", "CSC100Y9", "CSC100Y2", "CSC100Y4", "CSC100Y6", "CSC100Y7", "CSC100Y8"};
 
-        ArrayList<Course> courseArrayList = new ArrayList<>();
+        ArrayList<CourseChoice> courseArrayList = new ArrayList<>();
         for (String courseName: courseList) {
-            courseArrayList.add(returnCourseFromName(courseName));
+            courseArrayList.add(returnCourseChoiceFromName(courseName));
         }
 
         Student student = new Student();
@@ -101,11 +102,11 @@ public class TestStudentCourseSorting {
         student.plannedYCourses = new ArrayList<>();
         student.sortAllCourseLists();
 
-        List<Course> actualCourseArrayList = student.plannedFCourses;
+        List<CourseChoice> actualCourseArrayList = student.plannedFCourses;
         String[] actualCourseList = new String[actualCourseArrayList.size()];
 
         for (int i = 0; i < actualCourseArrayList.size(); i++) {
-            actualCourseList[i] = actualCourseArrayList.get(i).getCode();
+            actualCourseList[i] = actualCourseArrayList.get(i).course.getCode();
         }
 
         Assert.assertEquals(expectedCourseList, actualCourseList);
@@ -116,9 +117,9 @@ public class TestStudentCourseSorting {
         String[] courseList = {"AAA100Y1", "CSC100Y1", "BCB100Y1", "LOL100Y1", "HUH100Y1", "DOG100Y1", "CAT100Y1", "JOB100Y1", "CAD100Y1"};
         String[] expectedCourseList = {"AAA100Y1", "BCB100Y1", "CAD100Y1", "CAT100Y1", "CSC100Y1", "DOG100Y1", "HUH100Y1", "JOB100Y1", "LOL100Y1"};
 
-        ArrayList<Course> courseArrayList = new ArrayList<>();
+        ArrayList<CourseChoice> courseArrayList = new ArrayList<>();
         for (String courseName: courseList) {
-            courseArrayList.add(returnCourseFromName(courseName));
+            courseArrayList.add(returnCourseChoiceFromName(courseName));
         }
 
         Student student = new Student();
@@ -127,11 +128,11 @@ public class TestStudentCourseSorting {
         student.plannedYCourses = new ArrayList<>();
         student.sortAllCourseLists();
 
-        List<Course> actualCourseArrayList = student.plannedFCourses;
+        List<CourseChoice> actualCourseArrayList = student.plannedFCourses;
         String[] actualCourseList = new String[actualCourseArrayList.size()];
 
         for (int i = 0; i < actualCourseArrayList.size(); i++) {
-            actualCourseList[i] = actualCourseArrayList.get(i).getCode();
+            actualCourseList[i] = actualCourseArrayList.get(i).course.getCode();
         }
 
         Assert.assertEquals(expectedCourseList, actualCourseList);
